Add files via upload
The project focused on developing a predictive model to forecast high booking rates for Airbnb listings using a variety of models, including logistic regression, decision trees, random forests, LASSO, and ensemble methods like bagging and boosting. The best-performing model was a gradient boosting machine (GBM) with an accuracy of 85.76% and an AUC of 89.62%. The team effectively handled data preprocessing, feature engineering, and hyperparameter tuning, but faced challenges in hyperparameter optimisation and feature selection.